pyproject.toml
[project]
name = "market-agent"
version = "0.1.0"
description = "Daily market digest agent for US + TW markets (info aggregation only, no trading calls)"
requires-python = ">=3.12"
authors = [
    { name = "Kevin Chen", email = "kevin.chen@jkos.com" },
]
dependencies = [
    # Core data sources
    "yfinance>=0.2.50",        # US + TW equities & ETFs (uses pandas DataFrames natively)
    "fredapi>=0.5.2",          # FRED macro indicators
    "twstock>=1.3.0",          # Taiwan-specific supplementary data (外資買賣超 etc.)
    "httpx>=0.27.0",           # NewsAPI requests (modern replacement for requests)

    # Agent runtime
    "claude-agent-sdk>=0.1.0", # Claude Agent SDK (single-turn + future MCP integration)

    # Data / config
    "pandas>=2.2.0",           # Comes in transitively via yfinance/fredapi anyway
    "pyyaml>=6.0.1",           # Read config/*.yaml
    "python-dotenv>=1.0.1",    # Load .env

    # UX
    "rich>=13.7.0",            # Nice console output for local runs
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6.0",
    "pyright>=1.1.380",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fetchers", "agent"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # line too long — handled by formatter

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
